refactor(sync): move the unit's cover work out of the orchestrator
The three methods that deal with a work unit's cover images — the
invalidation pass, the download, and the cover_path stamped onto each
emitted entry — move to CoverPreparer, a new sub-service under
services/library/. The orchestrator drops 82 counted lines, 972 to 890,
and no longer holds the ArtworkManager seam at all.

The destination is not services/artwork.py, which already owns cover
downloading, and the reason is the direction the dependency would run.
attach_unit_cover_paths decides which ROMs to fetch covers for in the
sync's own delta vocabulary — BIND_ROM_ID_KEY, an emitted entry, a
WorkUnit — none of which the artwork service knows, and the other two
methods exist only to bind the run's emit_progress and is_cancelling
into a call whose callee is the artwork service itself. The artwork
service takes both as parameters precisely because it owns neither, and
holding them would mean importing services/library/_state.py, which the
service-independence contract forbids.

A module inside services/library/ leaves that contract's entry set
unchanged, and none of the three methods held a confined seam or opened
a transaction, so the seam-owner, read-only and UoW-nesting gates are
untouched.

Method bodies move verbatim; the one edit inside them is the progress
callback's source, which now arrives through the config the way the
fetcher's and the reporter's do. Peer-called methods lose their leading
underscore, matching the package's other async peers.

The moved tests move with them, into tests/services/library/
test_cover_preparer.py. Every other test that stubbed _download_artwork
follows the attribute to its new holder; no assertion changes.
