+++ py_modules/services/library/service.py
@@ -11,6 +11,7 @@ finalisation and the ``roms``-derived callable queries,
 :class:`SessionBudgetMonitor` for Steam's renderer-heap budget,
 :class:`ShortcutLaunchResolver` for each ROM's launch facts,
 :class:`ChunkDispatcher` for one unit's emit → ack → commit round-trips,
+:class:`CoverPreparer` for a unit's covers,
 :class:`LocalLibraryReader` for what this device already recorded about the
 library. The façade itself only wires the
 pieces together and delegates — anything that touches RomM or mutates
@@ -25,6 +26,7 @@ from typing import TYPE_CHECKING, Any
 from lib.late_binding import LateBinding
 from services.library._state import CollectionMembership, LibrarySyncStateBox
 from services.library.chunk_dispatcher import ChunkDispatcher, ChunkDispatcherConfig
+from services.library.cover_preparer import CoverPreparer, CoverPreparerConfig
 from services.library.fetcher import LibraryFetcher, LibraryFetcherConfig
 from services.library.local_library_reader import LocalLibraryReader, LocalLibraryReaderConfig
 from services.library.reporter import SyncReporter, SyncReporterConfig
@@ -105,7 +107,9 @@ class LibraryService:
     :class:`SessionBudgetMonitor` (Steam's renderer-heap budget),
     :class:`ShortcutLaunchResolver` (each ROM's installed path + active
     emulator), :class:`ChunkDispatcher` (one unit's apply, emitted and
-    committed a chunk at a time), and :class:`LocalLibraryReader` (this
+    committed a chunk at a time), :class:`CoverPreparer` (a unit's covers,
+    refreshed and downloaded before its shortcuts are emitted), and
+    :class:`LocalLibraryReader` (this
     device's own record of the library, read back out of SQLite) over a
     single shared :class:`LibrarySyncStateBox`. The façade itself owns the box and
     exposes the callable surface; every implementation method lives on
@@ -192,6 +196,18 @@ class LibraryService:
             )
         )
 
+        # Sub-service: cover preparer — the run's whole artwork surface, bound to
+        # this run's progress and cancel signals. Constructed before the
+        # orchestrator, which asks it for each unit's covers before the delta is
+        # handed to the dispatcher.
+        self._cover_preparer = CoverPreparer(
+            config=CoverPreparerConfig(
+                artwork=config.artwork,
+                sync_state_box=self._box,
+                emit_progress=self._emit_progress_proxy,
+            )
+        )
+
         self._orchestrator = SyncOrchestrator(
             config=SyncOrchestratorConfig(
                 settings=config.settings,
@@ -205,11 +221,11 @@ class LibraryService:
                 sync_state_box=self._box,
                 fetcher=self._fetcher,
                 reporter=reporter_binding,
-                artwork=config.artwork,
                 shortcut_launch_resolver=self._shortcut_launch_resolver,
                 local_library_reader=self._local_library_reader,
                 session_budget=self._session_budget,
                 chunk_dispatcher=self._chunk_dispatcher,
+                cover_preparer=self._cover_preparer,
             )
         )
