+++ tests/services/library/test_cover_preparer.py
@@ -1,0 +1,275 @@
+"""Tests for CoverPreparer — one work unit's covers, made ready before the apply.
+
+The preparer is reached through the library façade
+(``plugin._sync_service._cover_preparer``) so every test drives the same
+instance the orchestrator holds, over the shared state box the cancel signal
+lands on.
+
+Two levels are covered. The **delegation** tests call the preparer directly
+against a mocked ``ArtworkManager`` and pin what the run binds into each call —
+the progress position, the unit label, and an ``is_cancelling`` closure that
+tracks the live sync state rather than a snapshot of it. The **invalidation
+pass** is driven end-to-end through ``SyncOrchestrator._do_sync_per_unit``
+against the real ``ArtworkService`` (real cover-cache file I/O under
+``tmp_path``) and the seeded ``FakeRommApi``, because what the pass is for is
+observable only where it lands: the ``cover_refreshes`` list on the unit's first
+``sync_apply_unit`` chunk, and the fingerprint persisted beside it.
+
+``_wait_for_unit_complete`` stands in for a frontend ``report_unit_results``
+callback no test exercises; ``_download_artwork`` stands in for the SteamGridDB
+pipeline wherever the test is about the refresh pass rather than the download.
+"""
+
+import asyncio
+from unittest.mock import AsyncMock, MagicMock
+
+import pytest
+
+from domain.sync_state import SyncState
+
+# conftest.py patches decky before this import
+from tests.services.library._helpers import _fake_wait_set_event, _seed_platform, _seed_rom_row, _use_fake_romm
+
+
+class TestDownloadArtworkDelegation:
+    """Tests for _download_artwork."""
+
+    @pytest.mark.asyncio
+    async def test_delegates_to_artwork_manager(self, plugin):
+        """When _artwork is bound, the call is forwarded with progress + cancel hooks."""
+        fake_download = AsyncMock(return_value={1: "/path/a.png", 2: "/path/b.png"})
+        plugin._sync_service._cover_preparer._artwork = MagicMock()
+        plugin._sync_service._cover_preparer._artwork.download_artwork = fake_download
+
+        roms = [{"id": 1, "name": "A"}, {"id": 2, "name": "B"}]
+        result = await plugin._sync_service._cover_preparer._download_artwork(
+            roms, progress_step=3, progress_total_steps=7
+        )
+
+        assert result == {1: "/path/a.png", 2: "/path/b.png"}
+        fake_download.assert_called_once()
+        call_kwargs = fake_download.call_args.kwargs
+        assert call_kwargs["progress_step"] == 3
+        assert call_kwargs["progress_total_steps"] == 7
+        # is_cancelling closure reflects the live sync_state.
+        is_cancelling = call_kwargs["is_cancelling"]
+        plugin._sync_service._box.sync_state = SyncState.RUNNING
+        assert is_cancelling() is False
+        plugin._sync_service._box.sync_state = SyncState.CANCELLING
+        assert is_cancelling() is True
+
+    @pytest.mark.asyncio
+    async def test_forwards_unit_label_to_artwork(self, plugin):
+        """The unit display name is threaded through as the cover-progress label."""
+        fake_download = AsyncMock(return_value={})
+        plugin._sync_service._cover_preparer._artwork = MagicMock()
+        plugin._sync_service._cover_preparer._artwork.download_artwork = fake_download
+
+        await plugin._sync_service._cover_preparer._download_artwork(
+            [{"id": 1, "name": "A"}], progress_step=1, progress_total_steps=1, label="Game Boy Advance"
+        )
+
+        assert fake_download.call_args.kwargs["label"] == "Game Boy Advance"
+
+
+class TestCoverRefreshPass:
+    """The #1386 cover-cache invalidation pass wired through the per-unit apply.
+
+    Drives the real ArtworkService (real cover-cache file I/O under tmp_path)
+    against the seeded FakeRommApi, and asserts the refresh list rides the
+    unit's first ``sync_apply_unit`` chunk while the fingerprints persist.
+    """
+
+    _OLD = "/cover/big.png?ts=2026-01-01 00:00:00"
+    _NEW = "/cover/big.png?ts=2026-07-11 12:00:00"
+
+    @staticmethod
+    def _apply_setup(plugin, fake_romm_api):
+        import decky
+
+        decky.emit.reset_mock()
+        plugin.loop = asyncio.get_event_loop()
+        _use_fake_romm(plugin, fake_romm_api)
+        plugin._sync_service._cover_preparer._download_artwork = AsyncMock(return_value={})
+        plugin._sync_service._chunk_dispatcher._wait_for_unit_complete = _fake_wait_set_event
+        plugin._sync_service._box.sync_state = SyncState.RUNNING
+        plugin._sync_service._box.current_sync_id = "run-cover"
+
+    @staticmethod
+    def _apply_unit_events():
+        import decky
+
+        return [c[0][1] for c in decky.emit.call_args_list if c[0][0] == "sync_apply_unit"]
+
+    @staticmethod
+    def _cache_file(plugin, rom_id):
+        from pathlib import Path
+
+        return Path(plugin._artwork_service._cover_cache_dir) / f"{rom_id}.png"
+
+    @pytest.mark.asyncio
+    async def test_changed_cover_on_delta_skipped_rom_rides_first_chunk(self, plugin, fake_romm_api):
+        # rom 10 is content-unchanged (delta-skipped: no shortcut emitted) but its
+        # server cover source changed. The pass re-downloads the cache, persists
+        # the fresh fingerprint, and the {rom_id, app_id} entry rides chunk 0 so
+        # the frontend re-applies the tile.
+        self._apply_setup(plugin, fake_romm_api)
+        _seed_platform(
+            fake_romm_api,
+            platform_id=1,
+            name="N64",
+            slug="n64",
+            roms=[{"id": 10, "name": "Keep", "fs_name": "keep.z64", "path_cover_large": self._NEW}],
+        )
+        fake_romm_api.download_payloads[f"cover:{self._NEW}"] = b"fresh cover bytes"
+        plugin.settings["enabled_platforms"] = {"1": True}
+        _seed_rom_row(
+            plugin, 10, app_id=1010, platform_slug="n64", name="Keep", fs_name="keep.z64", cover_source=self._OLD
+        )
+
+        await plugin._sync_service._orchestrator._do_sync_per_unit()
+
+        events = self._apply_unit_events()
+        assert len(events) == 1
+        assert events[0]["shortcuts"] == [], "the item stays delta-skipped — a cover change never re-applies it"
+        assert events[0]["cover_refreshes"] == [{"rom_id": 10, "app_id": 1010}]
+        # The cache file holds the fresh bytes and the fingerprint advanced.
+        assert self._cache_file(plugin, 10).read_bytes() == b"fresh cover bytes"
+        with plugin._uow as uow:
+            assert uow.roms.get(10).cover_source == self._NEW
+
+    @pytest.mark.asyncio
+    async def test_null_fingerprint_adopts_without_refresh_entry(self, plugin, fake_romm_api):
+        # A pre-#1386 row (fingerprint NULL) with an existing cache file adopts
+        # the fresh fingerprint silently: no download, no refresh entry.
+        self._apply_setup(plugin, fake_romm_api)
+        _seed_platform(
+            fake_romm_api,
+            platform_id=1,
+            name="N64",
+            slug="n64",
+            roms=[{"id": 10, "name": "Keep", "fs_name": "keep.z64", "path_cover_large": self._NEW}],
+        )
+        plugin.settings["enabled_platforms"] = {"1": True}
+        _seed_rom_row(plugin, 10, app_id=1010, platform_slug="n64", name="Keep", fs_name="keep.z64", cover_source=None)
+        cache = self._cache_file(plugin, 10)
+        cache.parent.mkdir(parents=True, exist_ok=True)
+        cache.write_bytes(b"pre-existing cache")
+
+        await plugin._sync_service._orchestrator._do_sync_per_unit()
+
+        events = self._apply_unit_events()
+        assert len(events) == 1
+        assert events[0]["cover_refreshes"] == []
+        assert cache.read_bytes() == b"pre-existing cache", "NULL-adopt never re-downloads"
+        assert all(name != "download_cover" for name, _a, _k in fake_romm_api.call_log)
+        with plugin._uow as uow:
+            assert uow.roms.get(10).cover_source == self._NEW
+
+    @pytest.mark.asyncio
+    async def test_refreshes_ride_only_the_first_chunk(self, plugin, fake_romm_api, monkeypatch):
+        # Four changed items at chunk size 2 → two chunks; rom 1's cover also
+        # changed. The refresh entry rides chunk 0 only; chunk 1 carries [].
+        from services.library import chunk_dispatcher
+
+        self._apply_setup(plugin, fake_romm_api)
+        monkeypatch.setattr(chunk_dispatcher, "_APPLY_CHUNK_SIZE", 2)
+        _seed_platform(
+            fake_romm_api,
+            platform_id=1,
+            name="N64",
+            slug="n64",
+            roms=[
+                {
+                    "id": i,
+                    "name": f"New {i}",
+                    "fs_name": f"g{i}.z64",
+                    **({"path_cover_large": self._NEW} if i == 1 else {}),
+                }
+                for i in range(1, 5)
+            ],
+        )
+        plugin.settings["enabled_platforms"] = {"1": True}
+        for i in range(1, 5):
+            # Old names → every item classifies "changed" and is emitted.
+            _seed_rom_row(
+                plugin,
+                i,
+                app_id=1000 + i,
+                platform_slug="n64",
+                name=f"Old {i}",
+                fs_name=f"g{i}.z64",
+                cover_source=self._OLD if i == 1 else None,
+            )
+
+        await plugin._sync_service._orchestrator._do_sync_per_unit()
+
+        events = self._apply_unit_events()
+        assert len(events) == 2
+        assert events[0]["cover_refreshes"] == [{"rom_id": 1, "app_id": 1001}]
+        assert events[1]["cover_refreshes"] == []
+
+    @pytest.mark.asyncio
+    async def test_headroom_clips_refresh_list_before_emit(self, plugin, fake_romm_api):
+        # A live RSS reading leaves headroom for exactly ONE transient cover after
+        # the (empty) chunk's own cost: two refreshes clip to one — never a pause.
+        from domain.session_budget import CLIFF_KB, COVER_TRANSIENT_KB
+
+        self._apply_setup(plugin, fake_romm_api)
+        # The run's FIRST chunk projects against the cliff; leave headroom for one
+        # cover plus half of another so the allowance floor-divides to exactly 1.
+        plugin._renderer_rss.rss_kb = CLIFF_KB - COVER_TRANSIENT_KB - COVER_TRANSIENT_KB // 2
+        _seed_platform(
+            fake_romm_api,
+            platform_id=1,
+            name="N64",
+            slug="n64",
+            roms=[
+                {"id": 1, "name": "A", "fs_name": "a.z64", "path_cover_large": "/a.png?ts=2026-07-11 12:00:00"},
+                {"id": 2, "name": "B", "fs_name": "b.z64", "path_cover_large": "/b.png?ts=2026-07-11 12:00:00"},
+            ],
+        )
+        plugin.settings["enabled_platforms"] = {"1": True}
+        _seed_rom_row(
+            plugin, 1, app_id=1001, platform_slug="n64", name="A", fs_name="a.z64", cover_source="/a.png?ts=old"
+        )
+        _seed_rom_row(
+            plugin, 2, app_id=1002, platform_slug="n64", name="B", fs_name="b.z64", cover_source="/b.png?ts=old"
+        )
+
+        await plugin._sync_service._orchestrator._do_sync_per_unit()
+
+        events = self._apply_unit_events()
+        assert len(events) == 1, "the refreshes must never pause the run"
+        assert events[0]["cover_refreshes"] == [{"rom_id": 1, "app_id": 1001}], "clipped to the headroom allowance"
+        # Both grid-side caches were still refreshed backend-side; only the
+        # in-session tile push was clipped.
+        with plugin._uow as uow:
+            assert uow.roms.get(1).cover_source == "/a.png?ts=2026-07-11 12:00:00"
+            assert uow.roms.get(2).cover_source == "/b.png?ts=2026-07-11 12:00:00"
+
+    # ── delegation ───────────────────────────────────────────────
+
+    @pytest.mark.asyncio
+    async def test_refresh_delegates_to_artwork_manager(self, plugin):
+        fake_refresh = AsyncMock(return_value=[{"rom_id": 1, "app_id": 10}])
+        plugin._sync_service._cover_preparer._artwork = MagicMock()
+        plugin._sync_service._cover_preparer._artwork.refresh_changed_covers = fake_refresh
+
+        registry = {"1": {"app_id": 10, "cover_source": "/old.png?ts=1"}}
+        result = await plugin._sync_service._cover_preparer.refresh_changed_covers(
+            [{"id": 1, "name": "A"}], registry, progress_step=3, progress_total_steps=7, label="N64"
+        )
+
+        assert result == [{"rom_id": 1, "app_id": 10}]
+        assert fake_refresh.call_args.args == ([{"id": 1, "name": "A"}], registry)
+        call_kwargs = fake_refresh.call_args.kwargs
+        assert call_kwargs["progress_step"] == 3
+        assert call_kwargs["progress_total_steps"] == 7
+        assert call_kwargs["label"] == "N64"
+        # is_cancelling closure reflects the live sync_state.
+        is_cancelling = call_kwargs["is_cancelling"]
+        plugin._sync_service._box.sync_state = SyncState.RUNNING
+        assert is_cancelling() is False
+        plugin._sync_service._box.sync_state = SyncState.CANCELLING
+        assert is_cancelling() is True
